Marvin: More powerful behaviours for agents, ported DefaultPass, some cleanups
Each behaviour is a now a seperate class. In difference to a behaviour it is able to
have a "cooldown", that is to wait some time before it can be enabled again.
In addition it's much cleaner to have a behaviour stay active using its own state.

For passing the cooldown waits until the ball has arrive at another robot or a
timeout is reached. Meanwhile the agent is using its other behaviours like normal.