Revise situation tests
It is no longer required to have a robot on the field whose color and id is matching
that of the keeper encoded in the situation.

When there are more friendly robots on the field than required by the
situation, a warning is printed instead of an error. The surplus robots do not
take positions but will be controlled by the coordinator when the game starts.

When the number of opponent robots differs from that encoded in the situation,
a warning gets printed.

Synchronisation: a strategy only waits for the other if that is running a
situation test as well. That is, the other strategy breaks the "Halt" referee
state unless it is ready. This is distinguished from a button-pressed Force by
a timelimit. There is no more synchronisation via Amun for non-beaming
situation tests. For repeating a situation with 2 active strategies, both have
to be reloaded.

When beaming is enabled, an indicator via Amun is set to signal the other
strategy to reset its coordinator and thus all agents.
Only the blue strategy has to be reloaded then.