Marvin: Catchball: Rework obstacles
If the robot is hunting the ball, only add the current ball as obstacle.
If the robot is going to block the ball and thus could move past the robot, correct the error of modeling the robot as a circle.
Replace the moveCorridor by a tiny one that is just enough to guarantee that the robot approaches from behind and doesn't hit the ball.