Simulator: Only kick the ball after/when it has collided with the robot
If the ball is moving then it will probably touch the real robot before the kick is completed.
This is given by the fact, that the ball can only trigger a kick, once it's distance to the robot's kicker is less then a few (x) millimeters.
With a ball moving at y m/s this leaves x/y ms for the robot to kick the ball before it touches the dribbler. As the kick plunger is heavier than the ball
and given the fact that a kick duration of about 5-7 ms yields the fastest possible kick speed, the kick mechanism takes more time to execute the kick
than the ball needs to touch the dribbler.
Thus it is a reasonable approximation that a kick only takes place after the ball and robot have touched.