Amun: Strategy: create dummy team info for replay if none is present
This is useful for replaying logs where we switched sides during the
game.
Since only the team info for the color we first played is present,
the second half of the game can't properly be replayed if the one new
team info frame in the middle of the log isn't played during the replay.