marvin: o/error: Don't update speed error if the world is simulated
This is useful, because it is common practice to disable transciever in the simulator for testing purposes, to make the robots stand still and watch thier MoveTo.
Whithout this change, they'll default to t/error, as they are clearly not capable of moving, which is not the behavior that was desired.