marvin: g/centerback: don't comapare apples with peaches
When handleling the unimportant applications, g/centerback used target_way and robot_way to determine the position where the robot should go to.
After the change to 2018 rules, you can (if you want to) do some calculations to adjust the way even further, so that way+2*robotRadius will result in touching robots.
This is done using divCornerFactor and mulCornerFactor. When doing so, one has to make sure to compare two refined or two crude ways, but not a mixture.
This commit archieves that by refining target_way, too.