u/defense: move result of calculateBallPosition to DefenseArea
u/d/calculateBallPosition is always related to Centerbacks. This is obvious when considering the parameters.
Previously, cBP did return a value that had to be further processed to be a valid CB position. To prevent strange float issues
the first call was with a slightly larger defense Area

However, this resulted in comparing apples and bananas in g/centerback, when ways that have been calculated with different sized defense Areas where
compared using normal less_than, greater_than. Instead of trieing to distinguish between these different types of way, this commit removes the HACK FOR FLOATS
and uses one size of the defense Area at a time.