glados: u/interval prevent the merge function from creating new elements
The first element was read from the array and later written back to it.
If the array is empty, the first element ist undefined. But writing
undefined to the first position in the array will set the array length
to 1, although the content ist still undefined. This caused problems
down the line.