glados: u/attack fix creation of NaN in ratePass
checkedBallRollTime can return -Infinity by design.
This will cause ballAtTime to output a ball with a speed containing NaN.
This propagates to the rating result of the function.
In lua, this did not cause a problem as math.min(1, NaN) will result in
1 (although math.min(NaN, 1) will result in -NaN) (note that it is the
other way round for the luajit interpreter compared to the lua
interpreter). This prevented the spread of NaN, which doesn't happen in
javascript.