c++: enfoce rule of three
In c++, using either copy-constructor, destructor or assignment operator usually makes the default implementiation for all of these useless.
This commit enforces rule of three on every class in ra/horus/amun, and deletes the default copy-constructor and default assignment op.
This commit is not to be seen as reasoning why a class should not be copieable. It simply markes everything as not to be copied.
If a copy costructor appears to be usefull in future, simply decide it if it is really a thing you want.

Additionally, this commit replaces every tab in src, excluding firmware,
with 4 spaces.