Horus: Logprocessor: use SeqLogFileReader instead of LogFileReader
Using a LogFilereader results in indexing the log before starting to
read. This behavior is not needed for log processing.
However, this change makes the old progress display impossible.
The total number of frames that is going to be processed is no longer known.
Instead, the new progress is based on the memory that has already been processed,
and is seperated by logfile.