Ra: Logfilehasher: Change from Namespace to Class
This commit changes Logfilehash from beeing a namespace to a class.
All previous functions are changed to static functions.

This changes allows me to use the logfilehasher in logfilewriter in the
future, where no SeqLogFileReader is present, by simply adding the
status as they come to the new logfilehasher instance.