ra/amun: move simulator pause state to amun
As both the ui and strategies can pause and unpause the simulator and
the state was held in the ui only, this would lead to bugs when
focussing/unfocussing windows while debugging a strategy.
Now, the whole state ist moved to amun, where a number of different
sources can pause and unpause the simulator, which is centrally handled.