Horus: Logprocessor: Do not calculate Hash externally
Originally, the logprocessor had to calculate the hash itself.
Therefore, it used a SeqLogFileReader to read HASHED_PACKAGES only to
be used for calculating the hash. Afterwards, the same packages
have to be read for reencoding the whole logfile with the new hash.

Now, LogFileWriter is capable of hashing on its own. If no external
hash is supplied, it will calculate the hash "on the fly" while writing
the log.

This allows to use this feature for rehashing in Logprocessor.
Ihe Logprocessor is still responsible for creating the unified hash
for the resulting log.