amun: add support for replay to amun
Previously, replay was handled in the logplayer directly by handing
status packages to a strategy. With the future integration of logfile
status packages and simulator status packages for partial simulator,
this becomes infeasible.
Therefore, replay has to be integrated into amun.
The ReplayBlocker and the BlockingQueuedConnection are used to ensure
that amun is not slower in processing status packages than the ui in
generating them. The ui will be blocked for a short amount of time if
amun is too slow.