Protobuf: wrap StatusStrategy messages to avoid extensive duplication
In the future, strategy_replay_blue and strategy_replay_yellow would
have been needed, but this would have introduced unnecessary
duplication.
The StatusStrategy messages are not displayed in logs and are thus not
necessary to keep backwards compatibility. They are kept temporarily for
now to avoid conflicts with annotated logs from timeline, but should be
deleted later.
All usages are adapted to the new wrapper message.