Ra/Horus: don't clear backlog after saving it
It was previously cleared because during backlog saving, no status
packages were routed to the backlogwriter to not accumulate a lot of
memory. Not clearing it would have thus resulted in long breaks in the
next backlog (if it was created fast enough).
These problems are solved in this commit.