amun: js_path.cpp: bind lifetime to embedding typescript object
In Lua, the path object is destroyed by the lua GC,
and a callback is called to destroy the native ressource as well.
In Typescript / Javascript, this not trivial.
As a strategy is not (regularly) creating and destroying paths,
but should prefer to recycle path objects, this is no real problem.
However, a Typescript object does get destroyed regularly, i.e. when a reload triggers.
This results in leaking memory as Typescript did not keep track of its path objects.

This commit uses QT's parent & child relation, to bind the path object to its typescript object.
As the child will be deleted when its parent gets destroyed, this makes
sure that no memory can leak when reloading the strategy.
If a strategy does call pathCreate regulary, it has to make sure on its
own to call pathDestroy to clean up the ressources as well.