Ra: add classes for virtual field setup
The virtual field is for the following situation:
In a tournament, there is often only a small patch of a field available
for testing. Although the test might only need this small place (lets
say only a test of the goal keeper), the strategy may need a full sized
geometry to work properly.

The solution is to set up a virtual field, meaning that the actual
available field gets transformed into a larger field, on which only a
small area can be driven on (limited by vision areas).