CI: Limit allowed load for ninja
Library builds use their own instance of ninja which is not coordinating
its resource usage with the main instance. Thus we end up with way more
compiler processes than cpu core are available. This just ends up
wasting memory or might even causing swapping which in effect causes the
build to grind to an halt.