base/debugger: properly show multiple stack traces
Is is mostly unavoidable that all exceptions will produce a stack trace,
even those that are caught from within the typescript code itself.
Therefore, there can be multiple stack traces in one frame.
Notably, this happens in the unit test framework.
Now, the stack traces are shown with a number to seperate them.