glados: use stucturedStackTrace
Without this change, V8 produces a flat stackTrace for compatibility.
A flat stackTrace is a string that has been printed until now.
As a preparation for an integrated sourcemap support, the information is
now beeing presented in the structured way that v8 offers if the script
is not interested in compatibility with other browsers.
As our strategy is not supposed to run in any browser, we can easily
drop these for a more structured stackTrace instad of parsing the
string for the necessary information.