t/base/cache: properly name cached functions
This change is made primarly for improved readability of stacktraces.
Previously, a cached funtion had only a methodName, and no functionName.
This results in a simplified entry in the stackTrace, where the methodName
is shown as if it were a functionName. After this change, it has both,
a functionName and an methodName, resulting in the much more understandable
notation "at cachedFunction as xxx" when appearing in a stacktrace