base: use correct location for prepareStackTrace
Although the old location suggests that it will be called before any other statement, that is not true.
Instead, all imports are resolved before any normal line will be interpreted.
So, in order to have the really first line in our strategy, it has to be exported to a new module
that is first in the order of imports and does not import any other module on its own