glados: additional assertion for faulty forcedChange messages
There is currently no need to use the information, but it might be
needed in the future if other pools are introduced (for example
midfield). Therefore, the information is kept. However, an extra
assertion is added to make wrong use more obvious (at a later point,
another assertion would have fired).