glados: t/a/volley: bound v_in to Constants.maxBallSpeed
The caluclations don't make any sense if the ball is way faster. We still decided to bound the input
velocity to Constants.maxBallSpeed to avoid a crash if the ball seems to be too fast.
As the ball will not be faster in real play (or if it is faster, the play will be stopped),
we assume it not to be faster than the maximum allowed speed.