amun: typescript.cpp: Do not call debugger for caught exceptions
Our unit tests are very slow if they call the debugger all the time
for any raised exception. We want detailled information about not
correctly working unnit tests, but we do not need them for every
assert_error.

This change required the amun.tryCatch Interface to change.
The new Interface does not allow the catch block to return a boolean,
indicating if it handled the error sufficiently, or if ra should print a
stackTrace and therefore report it to the user.
Instead, this boolean has to be static, not depending on the error.