amun: js_path: new path interface for javascript
With the future addition of a second pathfinding which has some shared
functions with the rrt path finding, but also some different functions,
the interface needed to change to ensure that the correct C++
pathfinding object is used in all calls.
Instead of returning just the C++ pointer to the path object, an object
is created on which all path functions lie. These functions have the C++
object as an additional data field, which can not be modified from the
javascript side.
In this way, it is ensured that the correct C++ path object is passed to
each function.

As all functions should be backwards compatible with older strategies,
the original protocol is kept. To avoid code duplication, macros are
used to generate two different wrappers around the path functions, one
for the legacy protocol and one for the new protocol.