Ra: do not set robot color if we are playing as both teams
Previously, in a test were we play both teams (and the team name is
registered twice in the game controller or refbox), all robots were made
to be blue.
Now, the robot colors are not modified if we play both teams.
This will happen in internal tests which use the game controller and not
just the simulator.