ra/logfile: check very old logfiles for plausibility
Previously, SeqLogFileReader would accept every log. As the very old logformat did not contain an identification
Everything could be such an old log.

This is bad because it needs to be known when changing ra/logopener:
The Logopener uses multiple functions to determine, how to open this file.
Due to this technical limitation, LogFileReader had to be last in that list,
to make sure everybody gets a chance to claim this log.

Now, SeqLogFileReader - and therefore LogFileReader aswell - checks for these very old logs
if the first packet can be decoded. If not, it's most likely not a logfile.