glados: u/attack: fix crash caused by disparity of receivesPass and checkedBallRollTime
Fixes #618.
The problem is caused by the fact that ballRollTime extends the cone
behind the ball, which means that if the robot is close to the field
lines and the ball is 'inside' the robot, receivesPass might return true
while checkedBallRollTime returns infinity, since the ball can not be
caught inside the field.
The solution is to move the inf check down, so that both branches use
it.