Ra: fix window title not updating when opening logs
While switchToWidgetConfiguration will cause a title update when it is
currently in Ra mode, when Horus mode is already active, the widget
configuration is not changed and the title is therefore not update.
Therefore, always set it when opening a log file. If Ra mode is active,
the title will be overwritten (with the same string).