seshat: logfilefinder: implemented
LogFileFinder will be used to find a logfile, given a logfile::Uid.
It offers multiple logfiles, from which the user has to chose.

This commit also contains the following commit:

timeline: logfilefiner: avoid using fu2

While the old implementation was slightly less racy (due to the fact it kept all filedescriptors
open while the user was asked), and slightly more easy to add new functionality (as everything
just needed to fullfill the fu2::unique_function), it had a major issue:

Logfilefinder has to live in the newly created seshat module,
as it will create a StatusSource in the end. This makes it very annoying
to handle the list of functions propery, one has to send a message to
the UI, and wait for the response. The UI now has to send a message
"nevermind" back to seshat if the user decided not to open any of these
logs, or if the user decided for one log, it's index would have to be
used. All of these steps seems very unstable, and seem to invite a lot
new race conditions, for example if the user erroniously starts the
process twice.

Instead, the new approch works by using an universion ressource
identifier, which is right now just a local path, but can easily
modified. Instead of keeping the filedescriptors open, we close them
and reopen the logfile as soon as the user has selected his wish.

While this is racy in the sense that a user can rename the file while
this is in progress, and will then see an error message or a wrong log
after selecting the now outdated path in the UI, this was considered ok.
It also simplifies the abort case: If the user doesn't like any of the
offered logs, there's no need for any cleanup to be sent by the UI.

Also, getting rid of fu2 makes it easier to maintain our project.