glados: t/a/striker: add second hysteresis for choosing the pass dest
The first hysteresis only applies when the current robot will get the
pass. This hysteresis is for all other cases to ensure that the striker
does not flicker.
Also move the code for moveDest modification up to ensure that it is
properly visualized in the block that was previously above it.