glados: gr/midfield: only use passinfo pos when accepting pass
The idea behind this code probably was that accepting robots can get to
their destination.
The problem was that striker and midfield can send different
passsuggestions from the same robot, which, in conjunction with
overlapping striker and midfield zones, lead to a cycle of switching
that robot between midfield and striker (simplified situation).