Amun Processor: Use identical coordinate systems for current and speed command
The robot will compare it's target speed and the current speed (as seen by the cameras)
in the local coordinate form. Currently the conversion from global to local coordinate
system uses two distinct robot orientations. However, this has the problem that the
robot will compare the two (not really comparable) values. This seems to cause issues when
the robot's orientation starts to change.