amun: tsc: Update compilation state before notifying strategy
Sending the success signal while m_state is still RENAMING causes
a race between the requestPause call and the state update protected
with locker.relock() . Update state before sending the notification to
permanently fix this race condition.
Note: It is still possible that the strategy startup fails because the compiler
is busy renaming trees at that moment.