glados: a/a/default: reenable sidestep
This is a porting and typeing error:
In marvin, this reads as follows: if relevantPassInfo and not acceptingPass then

As the type of acceptingPass is boolean|undefined, it was falsely assumed that only true and undefined are used.
In fact, it is - at least in this part of the condition - surely != undefined, makeing this condition always false.