gitlab-ci: use -fsanitize=address as a thrid step
running amuncli with -fsanitize=adress does not trigger false-positives, and can detect memory leaks with an acceptable slowdown.

One could argue about replacing our default build processs and using -fsanitize=adress all the this, but i decided to use a third step instead.
This allows us to have a "realistic" test seperated, allowing us to disable this step if we start having problems with false-positives or other side-effects of using -fsanitize=address.
Using -fsanitize=undefined was considered but disregarded, as -fsanitize=undefined does not change the return value, which makes it not easily usable in ci.
Also, V8 seems to use some UB regarding "invalid vptr", which does not seem to be our fault.