amun: strategy/ts: Always save the ArrayBufferAllcator with its Isolate
This quasi-reverts 736d05e6f3.

An Isolate does not take ownership of its ArrayBufferAllcator.
Hence it needs to be deleted manually.

For typescript, the allocator has to be deleted when the Isolate is
disposed since external debuggers the it to be able to connect.

It may be possible to delete it earlier, e.g. after Isolate construction
in other cases. Documentation is not clear on this issue.