V8: Fix build hang with current msys version
The depot_tools update script starts a bootstrap script which is provided as
a .bat file. For some reason the call to the script manually tries to convert from a
msys internal path to a regular windows path. The manual conversion seems to
break with the latest msys version. Luckily, msys is already able to automatically
translate the path, thus just call the .bat file directly.