BuildBullet: Fix build when switching branches
When switching to a branch without PATCH_COMMAND, CMake will normally
retrigger everything after the download step. Note that this does not in
any way undo the PATCH_COMMAND.  Therefore when switching back to a
branch with PATCH_COMMAND, CMake will rerun the patch even though it is
already applied. This will then resulting in a patch application
failure.

Partially solve this by introducing an additional stage before the
download command that depends on the patch file.  This way cmake will
rerun the download step and end up with a clean build.  However, this
only works on the branch with the PATCH_COMMAND.

As the patch just fixes a compiler warning/error it is no problem to
still have the patch leftovers applied.