core/rng: Prevent broken RNG after seeding
The TAUSWORTHE rng calculations cannot produce random numbers when
seeded with 0. In that case only 0s will be emitted. As we use
microsecond timestamps for seeding the RNG, there is a small chance that
a seed of 0 is used. This will break the RNG and quite likely cause the
trajectory calculation to enter an infinite loop, effectively causing
the strategy to hang.

As we reseed every trajectory object for every robot on each strategy
iteration, after one hour of gameplay the probability for a hang is
already approx. 1:1000.