amun: js_path: Restore previous time to seed conversion behavior
Explicitly encode the "undefined" conversion behavior the was previously
used on 64bit platforms. This also properly handles negative numbers.
The only difference to the old behavior is for double value that cannot
be represented as int64_t.

The code that was emitted for the double to uint32_t conversion varies wildy
between 32 and 64bit platforms:

Tested with Compiler Explorer ( https://godbolt.org/ ):
> unsigned int conv(double s) {
>     return static_cast<unsigned int>(s);
> }

On 64bit platforms the following assembly is generated:
(x86-64 gcc 9.2, Options: "-O3")
> cvttsd2si       rax, xmm0
> ret
(see https://www.felixcloutier.com/x86/cvttsd2si for the opcode)
The call converts the double parameter to an 64bit integer which yields
a defined result as long as the double value fits into 64bits. The
latter is always the case for our seed values. The caller of the conv
function will then access the lower 32bits of rax, effectively truncating
the higher bits of the timestamp. Here the compiler interprets the
undefined behavior of out-of-value-range conversions to just mean the
fastest variant.

On 32bit platforms the following assembly is generated:
(x86-64 gcc 9.2, Options: "-O3 -m32 -msse2 -mfpmath=sse")
> maxArray(double):
>         movapd  xmm1, XMMWORD PTR .LC0
>         movq    xmm0, QWORD PTR [esp+4]
>         movapd  xmm2, xmm1
>         cmplepd xmm1, xmm0
>         andpd   xmm2, xmm1
>         pslld   xmm1, 31
>         subpd   xmm0, xmm2
>         cvttpd2dq       xmm0, xmm0
>         pxor    xmm0, xmm1
>         movd    eax, xmm0
>         ret
> .LC0:
>         .long   0
>         .long   1105199104
>         .long   0
>         .long   1105199104
The generated assembly effectively wraps the value range from [0,
2^32-1] to [-2^31, 2^31-1] using one-complement and reverses it after
conversion to an integer. This is equivalent to the following pseudocode
```
xmm0 = paramter
if (xmm0 >= 2^31) {
    xmm0 -= (double) 2^31
    xmm0 = (int32) xmm0
    xmm0 ^= (uint32) 2^31
} else {
    xmm0 = (int32) xmm0
}
return xmm0
```

The cvttpd2dp (double to int32 conversion) return 80000000H = 2^31 for
out of bound values. For a large timestamp this yields
```
xmm0 = (int32)(large double value) = (uint32)2^31
xmm0 ^= (uint32)2^31 = 0
```

FIX fabs