Fix LuaJIT crash on macOS Catalina
The call to lua_open failed and returned a null pointer that later on
caused the actual crash. In lua_open a mmap call is issued that seems to
fail on macOS Catalina. Luckily this problem is already (accidentially?)
fixed by upstream in LuaJIT 2.1.0-beta3. The patch just replaces
lj_alloc.c with the latest upstream version.