glados: tj/pathhelper: do add a short line for friendly passes
Even when not using CMA and therefore using the moving circle obstacle.

It turns out that the start of the pass is not very reliable regarding timing.
This results in blocking our own pass because the pass happens to be a bit delayed / a bit too early.

This fixed obstacle at the beginning makes sure that these inperfections are handled.