amun/simulator: the ball is now visible from multiple cameras in overlapping areas.
simball: Removed camera computations and moved them to simulator and
changed update to return a bool whether it is visible from the given
camera.
simulator: changed the return type of createVisionPacket to QList which
contains a packet for every camera that can see the ball and changed the
rest of the function accordingly. The computation of the camera position
is now consistent for both the ball and the calib part, the second of
which was hardcoded to 4 cameras and a camera height of 4.  The part
concerning the field and the calib portion of the function got
refactored to separate functions.