CMake: Properly trigger clobber step on library patch changes
The previous clobber step implementation was only run if the patch file
it depended on was newer than the timestamp of the last run of the
clobber step. Switching to a new patch file did work, however, reverting
to the stub patch failed to undo existing patch effects.

In addition the clobber step was triggered whenever the patch file was
touched, independent of whether the patch file actually changed.

The new implementation adds a custom command which copies the patch file
to an intermediate file which then serves as dependency target. Whenever
the source patch file _content_ changes the intermediate patch file is
modified. CMake will also delete the intermediate patch file when a
different source patch file is selected. This ensures that the library
is only rebuilt when necessary.