Fix patching of ExternalProject after switching to old branches
_make_ only ever rebuilds a target whenever it is missing or when one of
its dependencies are newer than the target. CMake (or Ninja) add one
additional check: A target will be rebuilt when the command line used to
generate it changed. This change detection only considers the command
itself but not the target's list of dependencies.

This leads to several restrictions for the implementation of the clobber
step (which triggers the download step whenever the used patch file for
an ExternalProject changes).

Adding a simple file dependency to a patch file to the download step is
not sufficient. This will miss cases where we just switch between two
existing patches. In that case it is likely that the stamp file for the
download step is newer than both patches and therefore the download step
won't be triggered by the patch dependency.

The solution to that is to include the patch file name into the command
of a make target. The solution chosen here is to have the clobber step
copy the patch file to an intermediate location on which the download
step then depends. This ensure that whenever the patch file or the patch
file name changes, the next make run will then update the intermediate
file. This in turn triggers the download step.

Rebuilding a dependency when only the patch timestamp but not its
content change is time consuming and wasteful. Therefore use the cmake
command copy_if_different to avoid rebuilds if nothing has changed.

When switching to an old branch without the clobber step, building the
code (parts of the problem also show up when skipping the build) and
then switching back the previous optimization will break. In the old
branch the intermediate patch file is left unchanged and therefore the
clobber step does not change the intermediate patch file and thus the
download step is not triggered again. The patch step, however, is run
again as its command has changed (cmake uses an empty command when no
patch step is specified). It will then try to apply the already applied
patch again and fails.

Another problem is that when switching back to the old branch any
changes made to the ExternalProject are NOT undone causing
non-reproducible builds. This is independent of whether the
copy_if_different optimization is used or not. As make and cmake don't
provide the means to trigger target rebuilds based on a changed
dependency list, a change to an existing target would be necessary.
Basically the only choice there would be the download command. It is
easy to modify whenever the source tarball changes but this becomes
impractical when a patch should be applied to an already used tarball.

Luckily it is possible to monkey patch the download step to add another
command. Just add some extra command to the download step. This ensures
that the download step command changes when switching to a non-clobber
branch and when switching back. Changes to the patch itself will be
picked up by the dependency on the intermediate patch copy.

The only alternative to monkey patching the download step would be the
following: Ensure that builds of the external project with and without
the clobber step don't interact. The easiest way to achieve this is to
change the build prefix for the external project when adding the clobber
step. The only downside here is that the build folder can over time
accumulate a build of the ExternalProject with and without the clobber
step. As this increase in file count and size would slow down the CI
further, the Gitlab CI cache key has also to be bumped.