glados: tj/pathhelper: do not avoid opponents when they are already gone
This commit changes the static line to fade out as soon as we are certain that the robot will have left the area.
While it is possible for an opponent to reenter this area for an opponent, we simply assume it to be unlikely.
Even if that happens, we can most likely find a new path as soon as that happens.