glados: t/a/placeball: Set initial offsets to undefined
The initial offset was wrong when the ball is not visible or near the
offset target on task startup. Updating it once with the real ball
position even if the ball is not currently visible is better than using
the robot position since the ball at least keeps its last visible
position.