logfile: backlogwriter: remove remainder
This remainder is b****. QT-docu tells you that you have to supply an index firstIndex() <= i <= lastIndex() to at
This is only an issue for Debug-Builds, as asking at(0) will raise an qAssert.
In Release, a lot of things magicly work together (esp: the cache is always full if firstIndex != 0)
to make sure logs are correct.