glados: t/s/duel agressive positioning
Duel tries to get between the ball and the goal in certain cases.
With the broader goal this is not helpful when the opponent is close
to the goal because they can just shoot next to the duelist.
Duel now checks how much of the goal would be defended with its positioning.
If the defensive position does not cover at least 70% of the goal,
the duelist attacks the opponent robot instead.