base/ball: check if robot is near the ball while ball accelerate extremely
In some cases during robo cup if a ball crosses a line it was slowed
down. That is why predict shot assumed the ball does not arrive.
Now: If a ball accelerate extremely we also check if a robot is in the
near. If not we update the ball speed correcly.