glados: a/b/behavior: Change return type of check to Behavior | undefined
To implement the new objective system, behaviors need to be able to
return other behaviors during check time. Using deferred behaviors is not
possible since the returned behaviors need to be instanced (to make it
possible to check them, because the are in a priority list), while
running a deferred behavior requires a Behavior constructor.