glados: a/b/agent: Move active behavior condition for MA parameters
This is needed to allow implementing a behavior composite class.

Such a class would contain a list of sub-behaviors and return one of
them in check, which will then be stored in `this._activeBehavior`.
However, the active behavior is not a member of `this._behaviors`
directly (since it is stored inside the composite class instead). Thus,
the old equality check would never be true.