Glados: tr/attackratio: passes do need two attackers
Previously in this situation, we would simply not pass.
With a recent change, we would pass to the defender as a piggy might suggest a pass
After requesting a pool-change, this would trigger 'MainAttacker is defender' aswell as a/d/handleball
, dropping the pass and changing back to the MA beeing an attacker.

To break that cycle, we increase the attackRatio in this case to two. The incrased attackRatio is no problem,
as the prefered robot will be the robot recieving a pass,
and sending a passSuggestion implies that this robot is not necessary for defending