glados: t/a/acceptpass: Accept directly adressed pass from passInfoTable
If there are more undefined passes in the passInfoTable the code should
throw an exeption. If a move uses undefined passes and handles them
correctly an error would occure for all robots with specific passes in the same
passInfoTable. To prevent this the robot searches for a pass exclusive
for him and ignores other undefined passes.