glados: t/a/placeball.ts: Fixed two attackers in placeBall sending placingRobot
In the first frame of placeball two robots try to place the ball.  The
current mainAttacker and the shooter that is going to be the
mainAttacker in the next frame. They both send placingRobot, which is an
exclusive role, leading to a crash. To prevent this the message is sent
only after the first frame.