glados:t/at/placeball: Fixed crash
To prevent using a bad default value for the offsets, the calculation
is delayed until the robot is close enough to calculate a good default
value or _calculateOffsets can calculate the offsets.

With modifying WAIT_FOR_BALL_STOP the offset values are reset every
frame till these offsets can be used.

This refers to issue #707, where a bad init value caused problems.
b02de2cbb solves this by avoiding to calculate an default offset,
which leads to rare situation where the offsets are undefined.