Ra: l/combinedlogwriter: Do not save Status with UIResponses
Saving empty Status with time only (and maybe no usable time at all) is bad.
Checking for "is any field set apart from pure_ui_response and time" is hard to impossible
due to design decisions in protobuf (there could be data that is no longer or not yet known to this instance
of Status), so we decided to disallow mixing UI-Response Packets and other Packets.
This also ensures that LogfileHashing - done by the logfilewriter - will never see these UIResponses