amun: simulator: increase the reported ball area
The data being reported by the simulator was around 5 pixel, which is
basically useless since it is an integer and therefore basically does
not vary.

The scale can in principle be chosen arbitrarily, since it directly
reflects the cameras resolution. A value was chosen to make it similar
to the RoboCup 2019