amun: trajectorypath: use zoned obstacle intersection instead of distance
The standard sampler cares about two intersection states:
Trajectories that fully intersect an obstacle are not allowed
Trajectories that come close to an obstacle are penalized to keep some
distance from obstacles if possible (but still allow driving close to
them if necessary).

Up until now, this has been done by computing the minimum obstacle
distance and using that. But the sqrt computation involved is
computationally complex and takes a long time.
Therefore, it is now replaced by a squared distance computation with
checks for two different obstacle "radii", the normal obstacle radius
and the radius extended by the "near obstacle" radius.

This commit should not change anything functionally, just improve the
performance.