base: vector: replace normalize() with non-modifying normliazed() function
The normalize function is dangerous since it can modify arguments for
example given as arguments if care is not taken. Therefore, it is often
seen following a .copy(). If this is forgotten, the arguemnt can be
modified without knowledge, leaving a different vector to the caller
when the function returns.

An example of this can be seen in t/a/movetoball (changed in this diff),
which would modify the viewDir argument (definitely a bug)